+++ disable-search.php
@@ -83,6 +83,8 @@ class c2c_DisableSearch {
 		add_filter( 'get_search_form', array( __CLASS__, 'get_search_form' ), 999 );
 
 		add_action( 'admin_bar_menu',  array( __CLASS__, 'admin_bar_menu' ), 11 );
+
+		add_filter( 'disable_wpseo_json_ld_search', '__return_true' );
 	}
 
 	/**