+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _(in-progress)_
 * Change: Note compatibility through WP 5.4+
+* Change: Update links to coffee2code.com to be HTTPS
 
 ## 1.7.2 _(2019-12-12)_
 * Change: Note compatibility through WP 5.3+