+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## _(in-progress)_
 * New: Disable output of `SearchAction` in SEO schema by Yoast SEO. Props @galengidman.
-* New: Add TODO.md and move existing TODO list from top of main plugin file into it
+* New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add to it)
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS